simplify slot resolution - re-enable dynamic slot names
