+++ src/compiler/index.js
@@ -1,3 +1,3 @@
 export * from './compile'
 export * from './transclude'
-export * from './scan-slots'
+export * from './resolve-slots'
