+++ src/util/env.js
@@ -8,12 +8,6 @@ export const inBrowser =
   typeof window !== 'undefined' &&
   Object.prototype.toString.call(window) !== '[object Object]'
 
-// Check if the browser supports native <template>.
-export const hasNativeTemplate = (function () {
-  var t = document.createElement('template')
-  return t.content && t.content.nodeType === 11
-})()
-
 // detect devtools
 export const devtools = inBrowser && window.__VUE_DEVTOOLS_GLOBAL_HOOK__
